# api.yml
openapi: "3.0.0"
info:
  title: Fullstack Challenge 20201209
  description: CRUD Users
  version: '0.1'
  title: "Users"
paths:
  /login: 
    post: 
      summary: 'Login user'
      description: 'Provide login form'
      tags: 
        - "Auth"
      responses:
        '200':
          description: 'Ok. Log in'
        '400':
          description: 'Invalid password'
        '404':
          description: 'Not found user'
  /register:
    post:
      summary: 'Register user'
      description: 'Provide a registration form'
      tags: 
        - 'Auth'
      responses:
        '200':
          description: 'Ok, user registrated'
        '400':
          description: 'Registration failed'
...